refactor: 그에 따른 변경사항
