+++ src/main/java/com/gloddy/server/auth/application/AuthService.java
@@ -8,7 +8,7 @@ import com.gloddy.server.auth.domain.dto.AuthRequest;
 import com.gloddy.server.auth.domain.dto.AuthResponse;
 import com.gloddy.server.core.error.handler.errorCode.ErrorCode;
 import com.gloddy.server.core.error.handler.exception.UserBusinessException;
-import com.gloddy.server.core.event.user.UserCreateEvent;
+import com.gloddy.server.user.event.UserCreateEvent;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
