+++ src/main/java/com/gloddy/server/auth/domain/User.java
@@ -5,15 +5,13 @@ import com.gloddy.server.auth.domain.vo.kind.Gender;
 import com.gloddy.server.auth.domain.vo.kind.Personality;
 import com.gloddy.server.core.converter.EnumArrayConverter;
 import com.gloddy.server.core.entity.common.BaseTimeEntity;
-import com.gloddy.server.core.event.reliability.ReliabilityScoreUpdateEvent;
-import com.gloddy.server.core.utils.event.GroupParticipateEvent;
+import com.gloddy.server.core.event.GroupParticipateEvent;
 import com.gloddy.server.praise.domain.Praise;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.group.domain.dto.GroupRequest;
 import com.gloddy.server.group.domain.handler.GroupCommandHandler;
 import com.gloddy.server.group.domain.service.GroupFactory;
 import com.gloddy.server.reliability.domain.Reliability;
-import com.gloddy.server.reliability.domain.vo.ScoreType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -116,7 +114,6 @@ public class User extends BaseTimeEntity {
         Group group = groupCommandHandler.save(groupFactory.getGroupFrom(this, groupInfo));
 
         eventPublisher.publishEvent(new GroupParticipateEvent(this.getId(), group.getId()));
-        eventPublisher.publishEvent(new ReliabilityScoreUpdateEvent(this.getId(), ScoreType.Created_Group));
         return group;
     }
 }
