+++ src/main/java/com/gloddy/server/group_member/domain/service/strategy/praise/AbsencePraiseStrategy.java
@@ -1,6 +1,6 @@
 package com.gloddy.server.group_member.domain.service.strategy.praise;
 
-import com.gloddy.server.core.event.group_member.GroupMemberReceivePraiseEvent;
+import com.gloddy.server.group_member.event.GroupMemberReceivePraiseEvent;
 import com.gloddy.server.group_member.domain.GroupMember;
 import com.gloddy.server.praise.domain.vo.PraiseValue;
 import org.springframework.context.ApplicationEventPublisher;
