+++ src/test/java/com/gloddy/server/acceptance/group/CreateGroupTest.java
@@ -1,7 +1,7 @@
 package com.gloddy.server.acceptance.group;
 
 import com.gloddy.server.common.group.GroupApiTest;
-import com.gloddy.server.core.utils.event.GroupParticipateEvent;
+import com.gloddy.server.core.event.GroupParticipateEvent;
 import com.gloddy.server.group.domain.dto.GroupRequest;
 import com.gloddy.server.group.domain.Group;
 import com.gloddy.server.group_member.domain.GroupMember;
